review(elicitation): widen elicit-timeout setting to float
Allows sub-second timeouts via env var (e.g. for tests). Drops the
redundant float() casts at the call sites in server.write() and
ask_ollie now that the source is already a float.

+++ src/opik_mcp/config.py
@@ -81,7 +81,7 @@ class Settings(BaseSettings):
     # Timeout is treated as a deny ("no") — the safer default; the user can
     # always re-issue the call. 60s matches typical chat-UI attention spans
     # without holding the host's tool-call slot open indefinitely.
-    opik_mcp_elicit_timeout_seconds: int = 60
+    opik_mcp_elicit_timeout_seconds: float = 60.0
 
     @field_validator("opik_mcp_auto_approve", "opik_mcp_confirm_writes", mode="before")
     @classmethod
